Add new images and update experience section
Added new images `development.png` and `hacker.png` to the assets. Updated the `AppStrings` class to include detailed descriptions for Flutter Developer roles. Created a new `Experience` model to manage experience data and updated the `ExperienceItem` widget to use this model. Added a new `ExperienceSliverGrid` widget to display experiences in a grid format and updated the `LandingViewDesktopAboutTab` and `LandingViewMobileAboutTab` to use this new grid. Updated `assets.gen.dart` to include the new images.